Small bug fixes, quality of life improvements, restructuring of where objects are, and login change
