+++ src/main/java/gui/ClassProfile.java
@@ -10,6 +10,7 @@ import java.util.Arrays;
 import java.util.ArrayList;
 import java.lang.Math.*;
 import java.text.*;
+import objects.*;
 
 import db.CategoryService;
